Optimizations and Refactoring
 -Moved ClockInfo struct to base
 -MasterClock uses chrono for all time values
 -RefClock partial implementation
 -Moved addition of now_µs to base Msg
 -Eliminated extra step from Frame handling (frame_release)
 -Added counting of future frames to Frames stats
 -Removed Frame::sync_wait_consume() (not needed)
 -Removed unneeded include and streamlined MasterClock API
 -Removed unused Reel::next_frame()
 -Added "finished" functionality to FX and MajorPeak
 -Improved "silent" Frame creation
 -Simplified Frame::next_frame by creation of state_now()
 -Spooler::next_frame() returns playable/future frames
 -Desk decides how to handle playable or future frames

+++ include/base/pet.hpp
@@ -79,7 +79,7 @@ struct pet {
   }
 
   static constexpr Millis from_ms(int64_t ms) { return Millis(ms); }
-  static constexpr Nanos from_ns(uint64_t ns) { return Nanos(ns); }
+  static constexpr Nanos from_ns(uint64_t ns) { return Nanos(static_cast<int64_t>(ns)); }
   static constexpr Nanos negative(Nanos d) { return Nanos::zero() - d; }
 
   template <typename T = Nanos> static T now_monotonic() {
